[Enhancement] Optimize concurrent partition creation cross transaction
