+++ fe/fe-core/src/main/java/com/starrocks/service/FrontendServiceImpl.java
@@ -2313,7 +2313,7 @@ public class FrontendServiceImpl implements FrontendService.Iface {
         try {
             // creating partition names is ordered
             for (String partitionName : creatingPartitionNames) {
-                txnState.lockCreatePartition(partitionName);
+                olapTable.lockCreatePartition(partitionName);
             }
 
             // ingestion is top priority, if schema change or rollup is running, cancel it
@@ -2348,7 +2348,7 @@ public class FrontendServiceImpl implements FrontendService.Iface {
             return result;
         } finally {
             for (String partitionName : creatingPartitionNames) {
-                txnState.unlockCreatePartition(partitionName);
+                olapTable.unlockCreatePartition(partitionName);
             }
         }
