+++ fe/fe-core/src/main/java/com/starrocks/transaction/TransactionState.java
@@ -79,8 +79,6 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReentrantLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 import javax.annotation.Nullable;
 import javax.validation.constraints.NotNull;
@@ -343,7 +341,6 @@ public class TransactionState implements Writable {
     // Therefore, a snapshot of this information is maintained here.
     private ConcurrentMap<String, TOlapTablePartition> partitionNameToTPartition = Maps.newConcurrentMap();
     private ConcurrentMap<Long, TTabletLocation> tabletIdToTTabletLocation = Maps.newConcurrentMap();
-    private Map<String, Lock> createPartitionLocks = Maps.newHashMap();
 
     private final ReentrantReadWriteLock txnLock = new FairReentrantReadWriteLock();
 
@@ -1015,28 +1012,6 @@ public class TransactionState implements Writable {
     public void clearAutomaticPartitionSnapshot() {
         partitionNameToTPartition.clear();
         tabletIdToTTabletLocation.clear();
-    }
-
-    public void lockCreatePartition(String partitionName) {
-        Lock locker = null;
-        synchronized (createPartitionLocks) {
-            locker = createPartitionLocks.get(partitionName);
-            if (locker == null) {
-                locker = new ReentrantLock();
-                createPartitionLocks.put(partitionName, locker);
-            }
-        }
-        locker.lock();
-    }
-
-    public void unlockCreatePartition(String partitionName) {
-        Lock locker = null;
-        synchronized (createPartitionLocks) {
-            locker = createPartitionLocks.get(partitionName);
-        }
-        if (locker != null) {
-            locker.unlock();
-        }
     }
 
     @Override
